move convert tool into a folder
