+++ Cargo.toml
@@ -30,7 +30,7 @@ path = "bin/level/main.rs"
 
 [[bin]]
 name = "convert"
-path = "bin/convert.rs"
+path = "bin/convert/main.rs"
 
 [dependencies]
 # internals
